:rewind: revert `responses` refactoring
